Creating detect image and extract it into a folder
You can run this script without worrying same exist folder, because this script will generate new folder every time you run it.
Please pay attention on detect_face.py line 6, it must be as same as the path of this folder xml. Also, in line 35-42 it must be adjust the as the user computer.
I tried it using as default path as I can, but it always return an error, so please if you can solve this problem.

Oh, one more thing, this script is still return an error even it is already satisfied the goal of this method. Suggestion: I think there is a problem for closing cv2.

Run using: python detect_face.py [IMAGE-PATH]. If you still confused, I enclosed the picture how to run this file.